Resolve "edit charge simcard"

Closes #310

See merge request Mostafa.abedi/traap-android!346

+++ app/src/main/java/com/traap/traapapp/ui/fragments/ticket/ticketInfo/TicketInfoInteractor.java
@@ -7,7 +7,7 @@ import com.traap.traapapp.apiServices.model.getTicketInfo.GetTicketInfoResponse;
  */
 public interface TicketInfoInteractor
 {
-    void reservationRequest(OnFinishedTicketInfoListener listener,int transactionId);
+    void reservationRequest(OnFinishedTicketInfoListener listener,Long transactionId);
 
 
     interface OnFinishedTicketInfoListener {
